+++ stacks-signer/src/v0/signer.rs
@@ -410,6 +410,11 @@ impl Signer {
                         ),
                         SignerMessage::StateMachineUpdate(update) => self
                             .handle_state_machine_update(signer_public_key, update, received_time),
+                        SignerMessage::BlockPreCommit(hash) => self.handle_block_pre_commit(
+                            stacks_client,
+                            &StacksAddress::p2pkh(self.mainnet, signer_public_key),
+                            hash,
+                        ),
                         _ => {}
                     }
                 }
@@ -716,6 +721,30 @@ impl Signer {
         self.impl_send_block_response(block_response)
     }
 
+    /// Send a pre block commit message to signers to indicate that we will be signing the proposed block
+    fn send_block_pre_commit(&mut self, block_hash: Sha512Trunc256Sum) {
+        info!(
+            "{self}: Broadcasting a block pre-commit to stacks node: {block_hash:?}";
+        );
+        match self
+            .stackerdb
+            .send_message_with_retry(SignerMessage::BlockPreCommit(block_hash))
+        {
+            Ok(ack) => {
+                if !ack.accepted {
+                    warn!(
+                        "{self}: Block pre-commit not accepted by stacker-db: {:?}",
+                        ack.reason
+                    );
+                }
+                crate::monitoring::actions::increment_block_pre_commits_sent();
+            }
+            Err(e) => {
+                warn!("{self}: Failed to send block pre-commit to stacker-db: {e:?}",);
+            }
+        }
+    }
+
     /// Handle signer state update message
     fn handle_state_machine_update(
         &mut self,
@@ -744,6 +773,92 @@ impl Signer {
             self.stacks_address,
             version,
         );
+    }
+
+    /// Handle pre-commit message from another signer
+    fn handle_block_pre_commit(
+        &mut self,
+        stacks_client: &StacksClient,
+        stacker_address: &StacksAddress,
+        block_hash: &Sha512Trunc256Sum,
+    ) {
+        debug!(
+            "{self}: Received a pre-commit from signer ({stacker_address:?}) for block ({block_hash})",
+        );
+        let Some(mut block_info) = self.block_lookup_by_reward_cycle(block_hash) else {
+            debug!(
+                "{self}: Received block commit for a block we have not seen before. Ignoring..."
+            );
+            return;
+        };
+        if block_info.has_reached_consensus() {
+            debug!("{self}: Received block pre-commit for a block that is already marked as {}. Ignoring...", block_info.state);
+            return;
+        };
+        // Make sure the sender is part of our signing set
+        let is_valid_sender = self.signer_addresses.iter().any(|addr| {
+            // it only matters that the address hash bytes match
+            stacker_address.bytes() == addr.bytes()
+        });
+
+        if !is_valid_sender {
+            debug!("{self}: Receive a pre-commit message from an unknown sender {stacker_address:?}. Will not store.");
+            return;
+        }
+
+        // commit message is from a valid sender! store it
+        self.signer_db
+            .add_block_pre_commit(block_hash, stacker_address)
+            .unwrap_or_else(|_| panic!("{self}: Failed to save block pre-commit"));
+
+        // do we have enough pre-commits to reach consensus?
+        // i.e. is the threshold reached?
+        let committers = self
+            .signer_db
+            .get_block_pre_committers(block_hash)
+            .unwrap_or_else(|_| panic!("{self}: Failed to load block commits"));
+
+        let commit_weight = self.compute_signature_signing_weight(committers.iter());
+        let total_weight = self.compute_signature_total_weight();
+
+        let min_weight = NakamotoBlockHeader::compute_voting_weight_threshold(total_weight)
+            .unwrap_or_else(|_| {
+                panic!("{self}: Failed to compute threshold weight for {total_weight}")
+            });
+
+        if min_weight > commit_weight {
+            debug!(
+                "{self}: Not enough committed to block {block_hash} (have {commit_weight}, need at least {min_weight}/{total_weight})"
+            );
+            return;
+        }
+
+        // have enough commits, so maybe we should actually broadcast our signature...
+        if block_info.valid == Some(false) {
+            // We already marked this block as invalid. We should not do anything further as we do not change our votes on rejected blocks.
+            debug!(
+                "{self}: Enough committed to block {block_hash}, but we do not view the block as valid. Doing nothing."
+            );
+            return;
+        }
+        // It is only considered globally accepted IFF we receive a new block event confirming it OR see the chain tip of the node advance to it.
+        if let Err(e) = block_info.mark_locally_accepted(false) {
+            if !block_info.has_reached_consensus() {
+                warn!("{self}: Failed to mark block as locally accepted: {e:?}",);
+            } else {
+                block_info.signed_self.get_or_insert(get_epoch_time_secs());
+            }
+        }
+
+        self.signer_db
+            .insert_block(&block_info)
+            .unwrap_or_else(|e| self.handle_insert_block_error(e));
+        let block_response = self.create_block_acceptance(&block_info.block);
+        // have to save the signature _after_ the block info
+        if let Some(accepted) = block_response.as_block_accepted() {
+            self.handle_block_signature(stacks_client, accepted);
+        };
+        self.send_block_response(block_response);
     }
 
     /// Handle block proposal messages submitted to signers stackerdb
@@ -997,12 +1112,12 @@ impl Signer {
         None
     }
 
-    /// Handle the block validate ok response. Returns our block response if we have one
+    /// Handle the block validate ok response.
     fn handle_block_validate_ok(
         &mut self,
         stacks_client: &StacksClient,
         block_validate_ok: &BlockValidateOk,
-    ) -> Option<BlockResponse> {
+    ) {
         crate::monitoring::actions::increment_block_validation_responses(true);
         let signer_signature_hash = block_validate_ok.signer_signature_hash;
         if self
@@ -1016,11 +1131,11 @@ impl Signer {
         let Some(mut block_info) = self.block_lookup_by_reward_cycle(&signer_signature_hash) else {
             // We have not seen this block before. Why are we getting a response for it?
             debug!("{self}: Received a block validate response for a block we have not seen before. Ignoring...");
-            return None;
+            return;
         };
         if block_info.is_locally_finalized() {
             debug!("{self}: Received block validation for a block that is already marked as {}. Ignoring...", block_info.state);
-            return None;
+            return;
         }
 
         if let Some(block_response) =
@@ -1036,15 +1151,8 @@ impl Signer {
             self.signer_db
                 .insert_block(&block_info)
                 .unwrap_or_else(|e| self.handle_insert_block_error(e));
-            None
         } else {
-            if let Err(e) = block_info.mark_locally_accepted(false) {
-                if !block_info.has_reached_consensus() {
-                    warn!("{self}: Failed to mark block as locally accepted: {e:?}",);
-                    return None;
-                }
-                block_info.signed_self.get_or_insert(get_epoch_time_secs());
-            }
+            block_info.valid = Some(true);
             // Record the block validation time but do not consider stx transfers or boot contract calls
             block_info.validation_time_ms = if block_validate_ok.cost.is_zero() {
                 Some(0)
@@ -1055,19 +1163,19 @@ impl Signer {
             self.signer_db
                 .insert_block(&block_info)
                 .unwrap_or_else(|e| self.handle_insert_block_error(e));
-            let block_response = self.create_block_acceptance(&block_info.block);
+            self.send_block_pre_commit(signer_signature_hash);
             // have to save the signature _after_ the block info
-            self.handle_block_signature(stacks_client, block_response.as_block_accepted()?);
-            Some(block_response)
+            let address = self.stacks_address;
+            self.handle_block_pre_commit(stacks_client, &address, &signer_signature_hash);
         }
     }
 
-    /// Handle the block validate reject response. Returns our block response if we have one
+    /// Handle the block validate reject response.
     fn handle_block_validate_reject(
         &mut self,
         block_validate_reject: &BlockValidateReject,
         sortition_state: &mut Option<SortitionsView>,
-    ) -> Option<BlockResponse> {
+    ) {
         crate::monitoring::actions::increment_block_validation_responses(false);
         let signer_signature_hash = block_validate_reject.signer_signature_hash;
         if self
@@ -1080,16 +1188,16 @@ impl Signer {
         let Some(mut block_info) = self.block_lookup_by_reward_cycle(&signer_signature_hash) else {
             // We have not seen this block before. Why are we getting a response for it?
             debug!("{self}: Received a block validate response for a block we have not seen before. Ignoring...");
-            return None;
+            return;
         };
         if block_info.is_locally_finalized() {
             debug!("{self}: Received block validation for a block that is already marked as {}. Ignoring...", block_info.state);
-            return None;
+            return;
         }
         if let Err(e) = block_info.mark_locally_rejected() {
             if !block_info.has_reached_consensus() {
                 warn!("{self}: Failed to mark block as locally rejected: {e:?}",);
-                return None;
+                return;
             }
         }
         let block_rejection = BlockRejection::from_validate_rejection(
@@ -1110,7 +1218,8 @@ impl Signer {
             .insert_block(&block_info)
             .unwrap_or_else(|e| self.handle_insert_block_error(e));
         self.handle_block_rejection(&block_rejection, sortition_state);
-        Some(BlockResponse::Rejected(block_rejection))
+        let response = BlockResponse::Rejected(block_rejection);
+        self.send_block_response(response);
     }
 
     /// Handle the block validate response returned from our prior calls to submit a block for validation
@@ -1121,15 +1230,15 @@ impl Signer {
         sortition_state: &mut Option<SortitionsView>,
     ) {
         info!("{self}: Received a block validate response: {block_validate_response:?}");
-        let block_response = match block_validate_response {
+        match block_validate_response {
             BlockValidateResponse::Ok(block_validate_ok) => {
                 crate::monitoring::actions::record_block_validation_latency(
                     block_validate_ok.validation_time_ms,
                 );
-                self.handle_block_validate_ok(stacks_client, block_validate_ok)
+                self.handle_block_validate_ok(stacks_client, block_validate_ok);
             }
             BlockValidateResponse::Reject(block_validate_reject) => {
-                self.handle_block_validate_reject(block_validate_reject, sortition_state)
+                self.handle_block_validate_reject(block_validate_reject, sortition_state);
             }
         };
         // Remove this block validation from the pending table
@@ -1137,11 +1246,6 @@ impl Signer {
         self.signer_db
             .remove_pending_block_validation(&signer_sig_hash)
             .unwrap_or_else(|e| warn!("{self}: Failed to remove pending block validation: {e:?}"));
-
-        if let Some(response) = block_response {
-            self.impl_send_block_response(response);
-        };
-
         // Check if there is a pending block validation that we need to submit to the node
         self.check_pending_block_validations(stacks_client);
     }
@@ -1420,10 +1524,9 @@ impl Signer {
             return;
         };
 
+        let stacker_address = StacksAddress::p2pkh(self.mainnet, &public_key);
         // authenticate the signature -- it must be signed by one of the stacking set
         let is_valid_sig = self.signer_addresses.iter().any(|addr| {
-            let stacker_address = StacksAddress::p2pkh(self.mainnet, &public_key);
-
             // it only matters that the address hash bytes match
             stacker_address.bytes() == addr.bytes()
         });
@@ -1437,7 +1540,10 @@ impl Signer {
         self.signer_db
             .add_block_signature(block_hash, signature)
             .unwrap_or_else(|_| panic!("{self}: Failed to save block signature"));
-
+        // If this isn't our own signature, try treating it as a pre-commit in case the caller is running an outdated version
+        if stacker_address != self.stacks_address {
+            self.handle_block_pre_commit(stacks_client, &stacker_address, block_hash);
+        }
         // do we have enough signatures to broadcast?
         // i.e. is the threshold reached?
         let signatures = self